requirement.txt
# Core LLM + API
mistralai
openai  # optional fallback / compatibility

# RAG frameworks
langchain
langchain-community
llama-index

# Embeddings
sentence-transformers
transformers
torch

# Vector databases
faiss-cpu
chromadb

# Data handling
pandas
numpy

# Document loaders / parsing
pypdf
python-docx
unstructured

# Web/API
fastapi
uvicorn

# Utilities
tqdm
python-dotenv
requests

# Optional: evaluation / debugging
ragas
datasets

accelerate
bitsandbytes

pinecone-client
weaviate-client
qdrant-client

streamlit
gradio 

langchain-mistralai